backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
langchain==0.0.340
pinecone-client==2.2.4
python-dotenv==1.0.0
pydantic==2.5.0
# Use newer pandas that supports Python 3.13
pandas==2.2.0
numpy==2.0.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
redis==5.0.1
# Skip psycopg2-binary for now, use SQLite
# psycopg2-binary==2.9.9
pypdf2==3.0.1
python-docx==1.1.0
# Add these for better compatibility
wheel
setuptools